% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\name{as.sup}
\alias{as.sup}
\title{Create a sup object.}
\usage{
as.sup(v, u = 0)
}
\arguments{
\item{v}{either a vector of length 2, in which case it is returned
unchanged, or a vector of length 1, in which case \code{u} (which defaults
to zero) is taken to be the uncertainty.}

\item{u}{a numerical value indicating the uncertainty in \code{v} (ignored
unless length(u) is 1).}
}
\value{
An object of class \code{sup}.
}
\description{
Create a sup object.
}
\examples{
as.sup(10)
as.sup(10, 1)
as.sup(c(10, 1))

}
