% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\name{\%/\%}
\alias{\%/\%}
\title{Divide two uncertain items}
\usage{
A \%/\% B
}
\arguments{
\item{A}{an uncertain value. This may be object of class \code{"u"}, created with \code{\link[=sup]{sup()}}, or a numeric vector of length 2 containing a value as the first
entry and its uncertainty as the second entry, or a numeric vector of length 1,
in which case the uncertainty is assumed to be zero.}

\item{B}{analogous to A.}
}
\value{
A two-element vector of class \code{"sup"}, which
contains result of the computation as the first entry,
and the uncertainty as the second.  (This class definition
means that printing the item will produce a a well-formated
display.)
}
\description{
Divide two uncertain items
}
\section{A note on operator precedence}{

The sup package does not follow conventional rules for operator
precedence. For example, in normal algebra, \code{1+2*3} is \code{7},
because the multiplication operator takes precedence over the
addition operator. By contrast, the sup operators work from left
to right, so that e.g. (in a case with zero uncertainty)
the value of \code{1 \%+\% 2 \%*\% 3} is \code{9}, because
the sum is done \emph{before} the multiplication. The use of parentheses
is highly encouraged to avoid confusion, e.g. write
\code{1 \%+\% (2 \%*\% 3)} to mimic conventional algebra.
}

\examples{
c(10, 1) \%/\% c(2, 0.5)
}
