% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\name{print.sup}
\alias{print.sup}
\title{Print an uncertain value}
\usage{
\method{print}{sup}(x, ...)
}
\arguments{
\item{x}{a \code{up}-class object to be printed}

\item{...}{ignored.}
}
\description{
Print an uncertain value
}
\examples{
print(sup(10, 1))
sup(10, 1)
c(10, 1) \%*\% c(100, 5)
}
