% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\docType{package}
\name{sup}
\alias{sup}
\title{sup: A Package for Simple Error Propagation}
\usage{
sup(v, u = 0)
}
\arguments{
\item{v}{either a vector of length 2, in which case it is returned
unchanged, or a vector of length 1, in which case \code{u} (which defaults
to zero) is taken to be the uncertainty.}

\item{u}{a numerical value indicating the uncertainty in \code{v} (ignored
unless length(u) is 1).}
}
\value{
An object of class \code{sup}.
}
\description{
The sup package provides functions for handling simple
error propagaation calculations. This is mainly handled
with binary operators such as \verb{\%+$} for addition.
}
\examples{
sup(10)
sup(10, 1)
sup(c(10, 1))

}
