% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\name{as.sup}
\alias{as.sup}
\title{Create a sup object.}
\usage{
as.sup(v, u = 0, x)
}
\arguments{
\item{v}{Either a vector of length 2, in which case it is returned
unchanged, or a vector of length 1, in which case \code{u} (which defaults
to zero) is taken to be the uncertainty. If missing, then \code{x}
must be supplied.}

\item{u}{A numerical value indicating the uncertainty in \code{v} (ignored
unless \code{length(u)} is 1, and ignored if \code{x} is supplied).}

\item{x}{A numerical vector of values. If this is given,
then the other arguments are ignored, and the return
value consists of the mean and the standard deviation
of x.}
}
\value{
An object of class \code{sup}.
}
\description{
Create a sup object.
}
\examples{
as.sup(10)       # 10 +- -
as.sup(10, 1)    # 10 +- 1
as.sup(c(10, 1)) # 10 +- 1
as.sup(x=1:10)   # 5.5 +- 3.02765

}
