% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\docType{package}
\name{sup}
\alias{sup}
\title{sup: A Package for Simple Error Propagation}
\description{
The sup package provides functions for handling simple
arithmetic operations for scalar quantities that are subject
to uncertainty. This is handled with the usual binary operators,
e.g. \verb{+} for addition, and with several unary functions, e.g.\preformatted{a <- sup(0.1, 0.01)
b <- sup(3, 0.1)
a * b                         # 3 +- 0
a \% b                         # 3 +- 0.4242641
sin(a)                        # 3 +- 0.4242641
}
}
\section{A note on left-to-right operator precedence}{


It is important to note that the sup operators work from
left to right, which is unlike the normal state of affairs
in R.  Thus, for example,\preformatted{as.sup(1, 0) + as.sup(2, 0) * as.sup(3, 0)
}

evaluates to \code{9 +- 0}, because the addition is done before
the multiplication, whereas conventonal R precedence
would suggest \code{7 +- 0}.  Parentheses are the solution
to this problem, e.g. writing\preformatted{as.sup(1, 0) + (as.sup(2, 0) * as.sup(3, 0))
}

in this case.

\strong{References}
\itemize{
\item Ku, Harry.
“Notes on the Use of Propagation of Error Formulas.”
Journal of Research of the National Bureau of Standards-C. Engineering and Instrumentation 70C, no. 4 (December 1966): 263–73.
\url{https://nvlpubs.nist.gov/nistpubs/jres/70C/jresv70Cn4p263_A1b.pdf}
\item Taylor, Barry N., and Chris E. Kuyatt.
“Guidelines for Evaluating and Expressing the Uncertainty of NIST Measurement Results.”
NIST Technical Note. Gaithersburg, MD, USA:
U.S. Department of Commerce Technology Administration: National Institute of Standards and Technology, 1994.
\url{https://www.nist.gov/pml/nist-technical-note-1297}
\item Ucar, Iñaki, Edzer Pebesma, and Arturo Azcorra.
“Measurement Errors in R.” The R Journal 10, no. 2 (2018): 549–57.
\url{https://doi.org/10.32614/RJ-2018-075}
\item Wikipedia.
“Propagation of Uncertainty.”
In Wikipedia, November 26, 2019.
\url{https://en.wikipedia.org/w/index.php?title=Propagation_of_ uncertainty&oldid=928003036}
}
}

